bedrock/nova/nova_reel/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
pillow>=10.0.0
boto3>=1.39.0
botocore>=1.39.0
python-dotenv>=1.0.0
